Add files via upload
In this repository, I have developed software that can take any output from the PROMIS Software and filter through its results. First, it mainly uses Python and may not operate on newer updates after 04/24. The main scope of the project is to help clean up the NJ Court System and how case files are tracked and handled outside of courts. However,  I know of several other states that use the same filing system and may benefit from a similar upgrade. This was an attempted universal approach that requires these steps before deployment: 

First Contact Database USA LLC and request access to the consumer data files they have available through one of their API services. Keep your credentials secure. Update the APISearch.py script with your new credentials.

Set up the appropriate Excel file end paths to parse data on to from your text file. If your court data reports are not presented on a text file you simply can copy and paste the info into one. Once the data report and the Excel files are in your folder,  configure the clean.py file to the appropriate scraping locations to match the format of your file. If you are in NJ or any other state using the Promises system this should be done for you already barring any major updates in the courts. 

Lastly, you can use the violation.py script to hyper-focus the results sent to the API. This is really an addition that was maybe unique to my use but I wanted a way to add an extra filter that you may not need. 

Any additional Meta tags can be pulled and adjusted using the metaTag.py script. 

A Second API is available for cross referencing for the most accurate information. Please contact the development team for more information.